requirements.txt
python-dotenv
flask
pymongo
flask_pymongo
numpy
pandas
scikit-learn
tensorflow
requests
markupsafe
flask_wtf